dump meets sydbox --dry-run (WIP)
* new configure flag --without-dry-run
* new command line switch -d, --dry-run
* shoebox learned "magic" subcommand to output sandboxing profile (WIP)

Previously sydbox used to ship a debug version with pinktrace functions
wrapped for primitive inspection of processes whereby sydbox-dump
produced a log of the traced processes. This changeset exposes the same
interface via --dry-run with small differences explained below.

Using --dry-run as opposed to sydbox-dump omits lower level ptrace()
details and focuses on inspection of the process rather than both the
process and the internal flow of sydbox.

"shoebox magic" may be used with the dump file to produce a sydbox
configuration profile which will run the inspected process under sydbox
whereby only the actions during the inspection run will be permitted.