replace ptrace with seccomp user notify
This is an attempt to replace ptrace with seccomp user notify.
The main loop is done. The double fork to transfer notify fd and exit
code is done.
The system call filters for all read, write, exec and network sandboxing
is done.
The connection to dereferencing the system call pointer arguments is
missing.

Signed-off-by: Alï Polatel <alip@exherbo.org>
Signed-off-by: Ali Polatel <alip@exherbo.org>