sydbox: import sha1 && json serializer
sha1dc is imported from git.git
head-commit:211eca0895794362184da2be2a2d812d070719d3
See src/sha1dc/LICENSE.txt for license information.
This is for checksumming execve() call arguments.

json serializer is imported from pg-to-json-serializer.git
head-commit:1de3c9acfbc1653b82038d5555e82e43c8d19762
Check GitHub project page or the LISENCE file at the root
of the git repository. GPL-3.

Note, only the json string escape utility of the json
serializer is imported. The rest of the functionality
is not needed in SydBox.

Signed-off-by: Alï Polatel <alip@exherbo.org>