sydbox: handle already existing rules gracefully when applying seccomp filters
Previously SydBox would abort when they tried to add a rule
and that rule is already present in the filter. As there is
no problem with filter having the rule already, SydBox may
just go ahead and add the remaining filter rules to the seccomp
bpf filter.

This was triggered by seccomp's socketcall() argument matching
for a bind() system call (a sub call in this case, e.g: on x86,
test with --arch x86 if your system supports it) and when SydBox
attempts to load an identical matches for the bind call ending
up in an EEXIST aborting the whole execution because it is
considered a fatal error.

Signed-off-by: Alï Polatel <alip@exherbo.org>