sydbox: fix problems with the SIGCHLD handler of sydbox --test seccomp (take 2)
This time I'm confident the problem is fixed. The SIGCONT
that the parent process sends to the child to continue with
the gettid call generates a SIGCHLD which SydBox wrongly
assumes as process exited because the handler is missing the
signal information.

This commit starts using sigaction and checks for the
termination information and only then notifies the
parent to go ahead with the seccomp user notification
testing.

Signed-off-by: Alï Polatel <alip@exherbo.org>