sydbox: honour EACCES return gracefully from rule add
EACCES means the rule addition failed because the default
action of the filter is the same as the action of the filter.
In such cases, we're almost always fine with the default action
and do not want to error out.

FIXME: This commit does not cover all such checks, but
almost all of them, there are a couple of direct
seccomp_rule_add calls that should be reviewed and
updated as necessary.

Signed-off-by: Alï Polatel <alip@exherbo.org>