sydbox: protect sandbox process from signals with a kernel-space only filter
For the kill() system call, there is a filter which
matches the process ID to SydBox's process and this filter
only permits the signal SIGCHLD to be sent. For any other
signal the system call returns ESRCH denoting the SydBox process
does not exist, rendering the sandbox process immune against
any signals from within the sandbox.

Functionally identical filters are applied for tkill and tgkill.
Since there's no way to match a process ID with the relatively
newer pidfd_send_signal system call, SydBox denies all signals
which have an action of term, core or stop with ESRCH.

This filter is unconditional from any option SydBox receives
and is there to increase the security of the sandbox.

Signed-off-by: Alï Polatel <alip@exherbo.org>