sydbox: simplify and fix open readonly code
Rather than trying to insert a ridicilously large
number of bpf filters (1024), generate the valid
open read only flags during compilation time which
reduces the number to 147 on a recent x86_64 Exherbo
GNU/Linux system with Linux-5.11.10.

This number is per-architecture (so on x86_64, we add
147*3 filters due to x86_64, x86 and x32) so we're still
way past the limit of 4096 bpf filters in total.

Hence this fixes errors with unable to load bpf filters at
startup when one of
-m core/sandbox/read:bpf
-m core/sandbox/read:allow
is given.

Signed-off-by: Alï Polatel <alip@exherbo.org>