sydbox: fix child with no parent sandboxing options inherit
This is a security fix, before a process which did not have
a parent known to SydBox could gain privileges to access
resources outside of the sandbox. With this change, SydBox
reparent them to the initial parent, makes sure this parent
never gets freed and if somehow the initial parent is gone
SydBox kills the process in question.

Signed-off-by: Alï Polatel <alip@exherbo.org>