TODO: add CaitSith part, thx eternaleye
Signed-off-by: Alï Polatel <alip@exherbo.org>
Poem-by: Lao Tse
Poem-from: Tao Te Ching
Poem-translated-by: Ursula K. Le Guin

No End

When government is lazy and informal
The people are kind and honest;
When government is efficient and severe
The people are discontented and deceitful.
Good fortune follows upon disaster;
Disaster lurks within good fortune;
Who can say how things will end?
Perhaps there is no end.
Honesty is ever deceived;
Kindness is ever seduced;
Men have been like this for a long time.
So the sage is firm but not cutting,
Pointed but not piercing,
Straight but not rigid,
Bright but not blinding.