libsyd: secure /proc interface (critical fix)
WARNING: This breaks the old insecure SydBox build, until
SydBox starts to use the new interface.

Two main changes:

1. Never access paths using a process ID
other than the two entry points:
- syd_proc_open -> /proc/pid
- syd_proc_fd_open -> /proc/pid/fd

Once the pid in these two calls is validated (e.g: by a
seccomp request identity) the rest of the libsyd
is secure to use.

2. Never call close() on a file descriptors
which is as an argument to a function.

Signed-off-by: Alï Polatel <alip@exherbo.org>